fix: 로그인 모달 수정사항 반영
